LayoutTests/platform/chromium/http/tests/inspector/console-xhr-logging-async-expected.txt

CONSOLE MESSAGE: line 14: Done.
CONSOLE MESSAGE: line 14: Done.
Tests that XMLHttpRequest Logging works when Enabled and doesn't show logs when Disabled for asynchronous XHRs.

Bug 79229
XHR finished loading: "http://127.0.0.1:8000/inspector/resources/xhr-exists.html". network-test.js:44
Done. console-xhr-logging-async.html:14
